fix(PR#1267 r4): reflector retry-fix narrow — ToBoundaryJson 移 Host/Adapter
修 r3 architect reject (CLAUDE 顶级约束 "外部协议 JSON 仅 Host/Adapter 边界"):

- 新建 src/Aevatar.Mainnet.Host.Api/Responses/ResponsesWebResultBoundaryJson.cs:
  独立 boundary JSON helper(System.Text.Json 渲染)
- src/platform/Aevatar.GAgentService.Abstractions/Responses/ResponsesWebResultJson.cs:
  移除 ToBoundaryJson + System.Text.Json 导入,只保留 FromLegacyValue /
  ToLegacyValue typed proto migration bridge
- src/platform/Aevatar.GAgentService.Projection/Queries/ResponsesAgentToolStateQueryReader.cs:
  不再调 ToBoundaryJson,readmodel snapshot 暴露 typed ResponsesWebToolResult
- ResponsesWebSubstituteToolJson.cs (Host/Api): 改用 ResponsesWebResultBoundaryJson
- 测试改名 ResponsesWebResultJsonTests → ResponsesWebResultMigrationTests
  (反映只测 typed proto migration bridge,boundary JSON 测试在 Host 层)

CLAUDE 条款引用 (verbatim):
"外部协议必须 JSON 时,仅在 Host/Adapter 边界做协议转换;进入应用/领域/
运行时层后恢复为 Protobuf。"

verify: dotnet build / test aevatar.slnx 全过(per CLAUDE rule 10)

⟦AI:AUTO-LOOP⟧
