+++ src/platform/Aevatar.GAgentService.Abstractions/Responses/ResponsesWebResultJson.cs
@@ -1,5 +1,3 @@
-using System.Text.Json;
-using Google.Protobuf;
 using Google.Protobuf.WellKnownTypes;
 
 namespace Aevatar.GAgentService.Abstractions.Responses;
@@ -8,13 +6,10 @@ namespace Aevatar.GAgentService.Abstractions.Responses;
 //   Old pattern: Responses Web fetch/search/error results moved through google.protobuf.Value
 //   as the normal internal contract, so actor state, cache, readmodels, and Host JSON each
 //   interpreted the same untyped payload shape.
-//   New principle: ResponsesWebToolResult is the typed internal contract; this boundary
-//   utility intentionally keeps the three transition duties together while the old Value
-//   surface remains only for legacy readmodel fallback and external JSON formatting.
+//   New principle: ResponsesWebToolResult is the typed internal contract; this migration
+//   utility only bridges legacy Value payloads to and from the typed contract.
 public static class ResponsesWebResultJson
 {
-    private static readonly JsonFormatter Formatter = new(new JsonFormatter.Settings(formatDefaultValues: false));
-
     public static ResponsesWebToolResult FromFetch(ResponsesWebFetchToolOutput output) =>
         new()
         {
@@ -47,7 +42,7 @@ public static class ResponsesWebResultJson
             return new ResponsesWebToolResult();
 
         if (value.KindCase != Value.KindOneofCase.StructValue)
-            return FromError("legacy_value_result", ToBoundaryJson(value));
+            return FromError("legacy_value_result", ReadValueAsString(value));
 
         var fields = value.StructValue.Fields;
         if (fields.TryGetValue("results", out var results) &&
@@ -87,61 +82,8 @@ public static class ResponsesWebResultJson
             });
         }
 
-        return FromError("legacy_value_result", ToBoundaryJson(value));
+        return FromError("legacy_value_result", "unsupported legacy result value");
     }
-
-    // Refactor (iter161-cluster-001 #1251-first):
-    //   Old pattern: Host-facing JSON was assembled from whichever untyped Value branch
-    //   happened to be present.
-    //   New principle: boundary JSON is rendered from typed ResponsesWebToolResult, with
-    //   legacy Value formatting confined to explicit fallback conversion.
-    public static string ToBoundaryJson(ResponsesWebToolResult? result)
-    {
-        if (result == null)
-            return "{}";
-
-        return result.ResultCase switch
-        {
-            ResponsesWebToolResult.ResultOneofCase.Fetch => ToBoundaryJson(result.Fetch),
-            ResponsesWebToolResult.ResultOneofCase.Search => ToBoundaryJson(result.Search),
-            ResponsesWebToolResult.ResultOneofCase.Error => ToBoundaryJson(result.Error),
-            _ => "{}",
-        };
-    }
-
-    public static string ToBoundaryJson(ResponsesWebFetchToolOutput? output)
-    {
-        var fields = new Dictionary<string, object?>
-        {
-            ["url"] = output?.Url ?? string.Empty,
-            ["status_code"] = output?.StatusCode ?? 0,
-            ["content_type"] = output?.ContentType ?? string.Empty,
-            ["content"] = output?.Content ?? string.Empty,
-        };
-        if (!string.IsNullOrWhiteSpace(output?.RedirectUrl))
-            fields["redirect_url"] = output.RedirectUrl;
-
-        return JsonSerializer.Serialize(fields);
-    }
-
-    public static string ToBoundaryJson(ResponsesWebSearchToolOutput? output) =>
-        JsonSerializer.Serialize(new
-        {
-            results = (output?.Results ?? Enumerable.Empty<ResponsesWebSearchResultItem>())
-                .Select(static item => new
-                {
-                    title = item.Title,
-                    url = item.Url,
-                    snippet = item.Snippet,
-                }),
-        });
-
-    public static string ToBoundaryJson(ResponsesWebToolError? error) =>
-        JsonSerializer.Serialize(new
-        {
-            error = error?.Code ?? string.Empty,
-            message = error?.Message ?? string.Empty,
-        });
 
     // Refactor (iter161-cluster-001 #1251-first):
     //   Old pattern: normal writes persisted stable Web result semantics as Value.
@@ -205,12 +147,6 @@ public static class ResponsesWebResultJson
         return value;
     }
 
-    private static string ToBoundaryJson(Value value)
-    {
-        using var document = JsonDocument.Parse(Formatter.Format(value));
-        return JsonSerializer.Serialize(document.RootElement);
-    }
-
     private static string ReadString(IDictionary<string, Value> fields, string name) =>
         fields.TryGetValue(name, out var value) ? ReadValueAsString(value) : string.Empty;
 
@@ -231,6 +167,6 @@ public static class ResponsesWebResultJson
             Value.KindOneofCase.NumberValue => value.NumberValue.ToString(System.Globalization.CultureInfo.InvariantCulture),
             Value.KindOneofCase.BoolValue => value.BoolValue ? "true" : "false",
             Value.KindOneofCase.NullValue => string.Empty,
-            _ => ToBoundaryJson(value),
+            _ => "unsupported legacy result value",
         };
 }
